+++ src/date-picker/wrapPicker.js
@@ -5,7 +5,7 @@
 
 import san from 'san';
 import {classCreator} from '../core/util/index';
-import localeReceiver from '../localeprovider/receiver';
+import localeReceiver from '../locale-provider/receiver';
 import dayjs from 'dayjs';
 import weekYear from 'dayjs/plugin/weekYear';
 import weekOfYear from 'dayjs/plugin/weekOfYear';