+++ src/pageheader/__tests__/index.spec.js
@@ -1,6 +1,0 @@
-/**
- * @file 组件 pageheader 测试文件
- * @author chenkai13 <chenkai13@baidu.com>
- */
-
-describe('pageheader', () => {});