+++ src/tabs/index.js
@@ -152,12 +152,10 @@ const Tabs = san.defineComponent({
         });
     },
     setActiveKey(key) {
-        const {activeKey, __activeKey__} = this.data.get();
-        const latestKey = activeKey || key;
-        if (__activeKey__ !== latestKey) {
-            this.data.set('__activeKey__', latestKey);
+        if (this.data.get('__activeKey__') !== key) {
+            this.data.set('__activeKey__', key);
             this.updateTab();
-            this.fire('change', latestKey);
+            this.fire('change', key);
         }
     },
     updateTab() {